+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `sandstone/WizardPanels` to prevent re-rendering of previous panel
+
 ## [2.0.0-rc.5] - 2021-07-22
 
 ### Added